requirements.txt
feedparser
facebook-sdk
django-web-rich-object
celery
